requirements.txt
absl-py==2.3.1
chex==0.1.91
flax==0.12.3
h5py==3.15.1
jax==0.8.2
jaxlib==0.8.2
jaxtyping==0.3.7
matplotlib==3.10.8
matplotlib-inline==0.2.1
nanobind==2.10.2
numpy==2.3.5
optax==0.2.6
orbax-checkpoint==0.11.31
orbax-export==0.0.8
scipy==1.16.3
